requirements.txt
fastapi
uvicorn
pandas
python-dotenv
groq
python-dateutil